requirements.txt
opencv-python
ultralytics
numpy
pyttsx3
torch
torchvision
transformers
Pillow
sentencepiece
protobuf
transformers
Pillow
sentencepiece
protobuf